+++ packages/react/src/components/Editor/CodeBlock/index.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { NodeViewContent, NodeViewWrapper } from '@tiptap/react';
+
+export const CodeBlock = ({
+  node: {
+    attrs: { language: defaultLanguage },
+  },
+  updateAttributes,
+  extension,
+}) => (
+  <NodeViewWrapper className="c-editor__code-block">
+    <select
+      contentEditable={false}
+      defaultValue={defaultLanguage}
+      onChange={(event) => updateAttributes({ language: event.target.value })}
+    >
+      <option value="null">auto</option>
+      <option disabled>—</option>
+      {extension.options.lowlight.listLanguages().map((lang, index) => (
+        <option key={index} value={lang}>
+          {lang}
+        </option>
+      ))}
+    </select>
+    <pre>
+      <NodeViewContent as="code" />
+    </pre>
+  </NodeViewWrapper>
+);